+++ src/components/anchor/anchor.test.tsx
@@ -1,0 +1,7 @@
+import { Anchor } from '.';
+
+describe('Anchor', () => {
+  it('is truthy', () => {
+    expect(Anchor).toBeTruthy();
+  });
+});