+++ src/components/link/link.stories.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { Link } from '.';
-
-export default {
-  title: 'Link',
-};
-
-export const Primary = () => <Link />;