+++ src/components/link/link.test.tsx
@@ -1,7 +1,0 @@
-import { Link } from '.';
-
-describe('Link', () => {
-  it('is truthy', () => {
-    expect(Link).toBeTruthy();
-  });
-});